The NPC PR 3/3 OK (#119)
* npc werk

* atleast it builds

* m_npc OK

* forgor format

* review

* format

* fix warns?

+++ include/libc64/malloc.h
@@ -9,7 +9,7 @@ void* malloc(size_t size);
 void* malloc_r(size_t size);
 void* realloc(void* ptr, size_t size);
 void free(void* ptr);
-void* calloc(s32 num, size_t size);
+void* calloc(size_t num, size_t size);
 void GetFreeArena(size_t* outMaxFree, size_t* outFree, size_t* outAlloc);
 s32 CheckArena(void);
 void MallocInit(void* heap, size_t size);
